Migrate search controller tests to questions_search controller test
