+++ test/integration/login_flow_test.rb
@@ -14,8 +14,9 @@ class LoginFlowTest < ActionDispatch::IntegrationTest
     assert_equal '/post?tags=question:question&template=question', request.fullpath
   end
 
+  # This test depend on question based search functionality
   test "search a question and go through post mechanism if question not found when user is not logged in" do
-    get '/search/questions/What'
+    get '/questions_search/What'
     follow_redirect!
     assert_equal '/post?tags=question:question&template=question&title=What&redirect=question', request.fullpath
